Add new capabilities
- module -T: improved import and regularization of tessellations of non-cubical
  domains, added [r]sel and anisofact to -morpho, added -ori transform ori for
  tesrs, added -transform crysym, improved -ori odf, improved -morpho cube and
  square, added anisotropy to -oridistrib, added special orientations
  (fcc_psc*), added val<1e-12 to -morphooptistop default value, fixed -morpho
  file to load only morpho-related information, added -morphooptiboundlower,
  -morphooptiboundupper, -orioptiboundlower, -orioptiboundupper, added
  tessellation keys modenb and groupnb, added -orioptilogval, fixed -stat*
  scale, added -morpho hex,hexh,hexv, added -transform setcellid, added
  anisofact for tesr cells, added non-deterministic seed generation (id -1),
  improved -morpho tocta, which can now be used with -periodic, improved svg
  output, made minor other fixes and improvements, cleaned code.
- module -M: fixed hex meshing of periodic tessellations (.per generated),
  fixed -dim 2 -interface cohesive, fixed -loadmesh on $Topology=0 meshes,
  added -statnode orientation, added -statelset anisofact, made minor other
  fixes, added -format etree,ntree (only for development).
- module -S: sped up ODF computation, added odf(weight=...), added odfindex(),
  made minor bug fixes.
- module -V: added -space ori and -ori options, added -showscalemin and
  -showscalemax, added Manim colors, made minor fixes.
- general: added tutorials, improved installation (added make userinstall),
  added developer tools and documentation (added make devinstall), fixed
  orilib.

+++ src/neper_t/net_ori/net_ori_.h
@@ -1,5 +1,5 @@
 /* This file is part of the Neper software package. */
-/* Copyright (C) 2003-2024, Romain Quey. */
+/* Copyright (C) 2003-2026, Romain Quey, CNRS. */
 /* See the COPYING file in the top-level directory. */
 
 #include<stdio.h>
@@ -10,7 +10,6 @@
 #include<time.h>
 #include<sys/time.h>
 
-#include"structIn_t.h"
 #include"ut.h"
 #include"neut_t.h"
 #include"orilib.h"
@@ -34,7 +33,7 @@ extern void net_ori_mtess_params (struct IN_T In, int level, struct MTESS MTess,
                                   char **pori, char **porisampling,
                                   char **porispread, char **pcrysym);
 
-extern int net_ori_mtess_randseed_rand (int *N, int *id, int *poly, int levelqty);
+extern unsigned long net_ori_mtess_randseed_rand (int *N, int *id, int *poly, int levelqty);
 
 extern void net_ori_orispread (char *orispread, struct SEEDSET *pSSet);
